requirements.txt
requests
gitpython
python-dotenv
jq
patch
google-generativeai